pyproject.toml
[tool.poetry]
name = "jwt_auth"
version = "0.1.0"
description = ""
authors = ["Sergei Sorochinskii <sergei.sorochinskii@lanars.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^22.1"
psycopg2-binary = "^2.9.3"
PyJWT = "^2.4.0"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
